WG1/pke-ROLLO/pyproject.toml
# Configuration file for poetry python manager
# Run poetry install & poetry run python rbc-lib.py 

[tool.poetry]
name = "rbc-lib"
version = "1.2"
description = "Rank-Based Cryptography Library"
authors = ["Nicolas Aragon", "Loïc Bidoux"]
license = "LGPL"

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
